% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_means_and_plot.R
\name{calculate_means_and_plot}
\alias{calculate_means_and_plot}
\title{Calculate Means and Plot Results}
\usage{
calculate_means_and_plot(samples_re)
}
\arguments{
\item{samples_re}{The list result from the \code{generate_samples} function.}
}
\value{
A list containing the means of the samples and a graph plotting each sample mean around the "true" mean.
}
\description{
Calculate Means and Plot Results
}
\examples{
re <- generate_samples(5, 10, 70, 3.3)
mean_re <- calculate_means_and_plot(re)
mean_re
}
