% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_values_and_plot_histogram.R
\name{count_values_and_plot_histogram}
\alias{count_values_and_plot_histogram}
\title{Count Values and Plot Histogram}
\usage{
count_values_and_plot_histogram(samples, specific_value, comparison)
}
\arguments{
\item{samples}{The list result from the \code{generate_samples} function.}

\item{specific_value}{A numeric value to compare against.}

\item{comparison}{A string, either "small" or "larger", specifying whether to count values smaller or larger than \code{specific_value}.}
}
\value{
A list of counts of values smaller or larger than the \code{specific_value} in each sample, and a histogram of the counts.
}
\description{
Count Values and Plot Histogram
}
\examples{
re <- generate_samples(5, 10, 70, 3.3)
samples_re <- re$samples
count_re <- count_values_and_plot_histogram(samples_re, 50, "smaller")
count_re
}
