% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_samples.R
\name{generate_samples}
\alias{generate_samples}
\title{Generate Samples from a Normal Distribution}
\usage{
generate_samples(sample_size, sample_num, mean, SD)
}
\arguments{
\item{sample_size}{An integer specifying the size of each sample.}

\item{sample_num}{An integer specifying the number of samples.}

\item{mean}{A numeric value specifying the mean of the normal distribution.}

\item{SD}{A numeric value specifying the standard deviation of the normal distribution.}
}
\value{
A list containing \code{sample_num} lists of \code{sample_size} values randomly generated from a normal distribution with specified \code{mean} and \code{SD}, and the \code{mean} value.
}
\description{
Generate Samples from a Normal Distribution
}
\examples{
re <- generate_samples(5, 10, 70, 3.3)
re
}
